Navbar + Home in another file
Bottom Navbar created + Home's screen moved to another file.
